Send `X-Frame-Options` header only for web pages
Sending the `X-Frame-Options` header for everything by default,
doesn't bring any value when it comes to providing clickjacking
protection.

Even if some developers may want to send this header for some
other content (e.g.: images) to block other websites from using
it, for the majority that won't be the case.

Ref: h5bp/server-configs-apache#8
     h5bp/server-configs-apache#9